Adding crypto node pool sensor (12)

+++ custom_components/sensor/cryptonodepool.py
@@ -98,7 +98,7 @@ class CryptoNodePoolSensor(Entity):
             .strftime('%Y-%m-%d %H:%M:%S')
         last_payment_date = datetime.datetime.fromtimestamp(int(self._ticker['payments'][1]))\
             .strftime('%Y-%m-%d %H:%M:%S')
-        last_payment_value = long(self._ticker['payments'][0].split(':')[2]) / 1000000000000
+        last_payment_value = int(self._ticker['payments'][0].split(':')[2]) / 1000000000000
         return {
             STAT_HASHES: self._ticker['stats']['hashes'],
             ATTR_ATTRIBUTION: CONF_ATTRIBUTION,
@@ -108,7 +108,7 @@ class CryptoNodePoolSensor(Entity):
             PAYMENT_THRESHOLD: self._ticker['stats']['thold'] / 1000000000000,
             INVALID_SHARES: self._ticker['stats']['invalid'],
             PAID: self._ticker['stats']['paid'],
-            HASHRATE: long(self._ticker['stats']['hashrate']) / 1000,
+            HASHRATE: int(self._ticker['stats']['hashrate']) / 1000,
             LAST_PAYMENT_DATE: last_payment_date,
            LAST_PAYMENT_VALUE: last_payment_value
         }
